Path traversal audit mode tests separated
